requirements.txt
# Flask Web Framework
flask>=3.1.2

# Environment variable management
python-dotenv>=1.2.1

# Google Gemini AI API
google-generativeai>=0.8.5

# RSS Feed Parser for Indeed job listings
feedparser>=6.0.12
